gh-pages: put latest tag in 'stable' dir too

+++ scripts/build-gh-pages.sh
@@ -4,6 +4,7 @@ set -e
 id=$(git rev-parse HEAD)
 base="./gh-pages/$id"
 latest="./gh-pages/latest"
+stable="./gh-pages/stable"
 topReadme="./gh-pages/README.md"
 tag=$(git describe --exact-match --tags HEAD 2>/dev/null || echo "")
 
@@ -16,12 +17,14 @@ cp -r "./demo" "$base/demo"
 cp -r "./docs" "$base/docs"
 cp -r "./api-docs" "$base/api-docs"
 
-if [ ! -z "$tag" ]; then
+if [ ! -z "$tag" ] && [[ $tag == v* ]]; then
   echo "Detected tag: $tag"
   tagloc="./gh-pages/$tag"
   rm -rf "$tagloc"
   # would be nicer as a symlink, but doesn't work on travis
   cp -r "./gh-pages/$id" "$tagloc"
+  rm -rf "$stable"
+  cp -r "./gh-pages/$id" "$stable"
 fi
 
 rm -rf "$latest"
